requirements.txt
as==0.1
azure-cognitiveservices-vision-face==0.5.0
azure-common==1.1.26
certifi==0.0.8
chardet==1.0.1
chardet2==2.0.3
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
imageio==2.9.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
msrest==0.6.21
networkx==2.5
np==1.0.2
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.1.0
pip-autoremove==0.9.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==0.11.1
requests-oauthlib==1.3.0
six==1.15.0
tifffile==2021.2.1
urllib3==1.26.4
Werkzeug==1.0.1